Sphere, H3DAPI - Fixed so that only one vbo id is used for all spheres, scaling is done before calling list.
The rest of the files, small changes to get rid of warning messages. Such as conversion warnings and such. Discovered a copy and paste error in SpiderMonkeyTypes that needed to be fixed (correction to SFVec3d type, it was the same as SFVec3f type before).

git-svn-id: https://www.h3d.org:8090/H3DAPI/trunk/H3DAPI@1635 610dfd11-2700-0410-9b68-9ae28eac5983

+++ src/NurbsTextureCoordinate.cpp
@@ -154,7 +154,8 @@ NurbsTextureCoordinate::NurbsTextureCoordinate(
   for( unsigned int i = 0; i < noWeights.size(); i++ ) {
     withWeights[ i * sizeOfVertex ] = (GLfloat)noWeights[i].x;
     withWeights[ i * sizeOfVertex + 1 ] = (GLfloat)noWeights[i].y;
-    if ( theWeights.size() == 0 || theWeights.size() < u_dimension * v_dimension ) {
+    if ( theWeights.size() == 0 ||
+         theWeights.size() < (unsigned int)( u_dimension * v_dimension ) ) {
       withWeights[ i * sizeOfVertex + 2 ] = (GLfloat)1.0;
     } else {
       withWeights[ i * sizeOfVertex + 2 ] = (GLfloat)theWeights[i];
